camera/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
black==24.8.0
blinker==1.8.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
detectron2==0.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
fvcore==0.1.5.post20221221
groq==0.11.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
hydra-core==1.3.2
idna==3.9
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
openai==1.45.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.3
portalocker==2.10.1
protobuf==5.28.1
psutil==6.0.0
py-cpuinfo==9.0.0
py-spy==0.3.14
pycocotools==2.0.8
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.93
ultralytics-thop==2.0.6
urllib3==2.2.3
Werkzeug==3.0.4
yacs==0.1.8
